Split thread_local_object.h and overload_manager.h
Splitting these will prevent circular dependencies when the Dispatcher
references thread-local overload state.
