+++ include/envoy/runtime/BUILD
@@ -17,7 +17,7 @@ envoy_cc_library(
     ],
     deps = [
         "//include/envoy/stats:stats_interface",
-        "//include/envoy/thread_local:thread_local_interface",
+        "//include/envoy/thread_local:thread_local_object",
         "//source/common/common:assert_lib",
         "//source/common/singleton:threadsafe_singleton",
         "@envoy_api//envoy/type/v3:pkg_cc_proto",
