+++ include/envoy/runtime/runtime.h
@@ -9,7 +9,7 @@
 
 #include "envoy/common/pure.h"
 #include "envoy/stats/store.h"
-#include "envoy/thread_local/thread_local.h"
+#include "envoy/thread_local/thread_local_object.h"
 #include "envoy/type/v3/percent.pb.h"
 
 #include "common/common/assert.h"
