+++ include/envoy/server/BUILD
@@ -127,7 +127,7 @@ envoy_cc_library(
         "//include/envoy/local_info:local_info_interface",
         "//include/envoy/runtime:runtime_interface",
         "//include/envoy/secret:secret_manager_interface",
-        "//include/envoy/server:overload_manager_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
         "//include/envoy/ssl:context_manager_interface",
         "//include/envoy/thread_local:thread_local_interface",
         "//include/envoy/tracing:http_tracer_interface",
@@ -151,8 +151,8 @@ envoy_cc_library(
     name = "worker_interface",
     hdrs = ["worker.h"],
     deps = [
-        ":overload_manager_interface",
         "//include/envoy/server:guarddog_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
     ],
 )
 
@@ -183,7 +183,7 @@ envoy_cc_library(
         "//include/envoy/local_info:local_info_interface",
         "//include/envoy/network:drain_decision_interface",
         "//include/envoy/runtime:runtime_interface",
-        "//include/envoy/server:overload_manager_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
         "//include/envoy/singleton:manager_interface",
         "//include/envoy/thread_local:thread_local_interface",
         "//include/envoy/tracing:http_tracer_interface",
@@ -207,8 +207,8 @@ envoy_cc_library(
         "//include/envoy/config:typed_config_interface",
         "//include/envoy/http:codes_interface",
         "//include/envoy/http:filter_interface",
-        "//include/envoy/server:overload_manager_interface",
         "//include/envoy/server:transport_socket_config_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
         "//include/envoy/singleton:manager_interface",
         "//include/envoy/thread_local:thread_local_interface",
         "//include/envoy/tracing:http_tracer_interface",
@@ -299,16 +299,6 @@ envoy_cc_library(
         "//include/envoy/api:api_interface",
         "//include/envoy/event:dispatcher_interface",
         "//include/envoy/protobuf:message_validator_interface",
-    ],
-)
-
-envoy_cc_library(
-    name = "overload_manager_interface",
-    hdrs = ["overload_manager.h"],
-    deps = [
-        "//include/envoy/event:timer_interface",
-        "//include/envoy/thread_local:thread_local_interface",
-        "//source/common/singleton:const_singleton",
     ],
 )
