+++ include/envoy/server/worker.h
@@ -3,7 +3,7 @@
 #include <functional>
 
 #include "envoy/server/guarddog.h"
-#include "envoy/server/overload_manager.h"
+#include "envoy/server/overload/overload_manager.h"
 
 namespace Envoy {
 namespace Server {
