+++ include/envoy/thread_local/BUILD
@@ -11,5 +11,15 @@ envoy_package()
 envoy_cc_library(
     name = "thread_local_interface",
     hdrs = ["thread_local.h"],
-    deps = ["//include/envoy/event:dispatcher_interface"],
+    deps = [
+        ":thread_local_object",
+        "//include/envoy/event:dispatcher_interface",
+        "//source/common/common:assert_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "thread_local_object",
+    hdrs = ["thread_local_object.h"],
+    deps = ["//source/common/common:assert_lib"],
 )
