+++ source/common/memory/BUILD
@@ -35,7 +35,7 @@ envoy_cc_library(
     deps = [
         ":utils_lib",
         "//include/envoy/event:dispatcher_interface",
-        "//include/envoy/server:overload_manager_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
         "//include/envoy/stats:stats_interface",
         "//source/common/stats:symbol_table_lib",
     ],
