+++ test/mocks/server/BUILD
@@ -161,7 +161,7 @@ envoy_cc_mock(
     srcs = ["overload_manager.cc"],
     hdrs = ["overload_manager.h"],
     deps = [
-        "//include/envoy/server:overload_manager_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
         "//test/mocks/event:event_mocks",
     ],
 )
@@ -207,7 +207,7 @@ envoy_cc_mock(
     hdrs = ["main.h"],
     deps = [
         "//include/envoy/server:configuration_interface",
-        "//include/envoy/server:overload_manager_interface",
+        "//include/envoy/server/overload:overload_manager_interface",
     ],
 )
